+++ R/expect-conditional.R
@@ -23,9 +23,7 @@
 #' )
 #'
 #' # Check that q1b has a value if and only if q1a %in% 1
-#' \dontrun{
-#' expect_base(q1b, q1a %in% 1, data = my_survey) # Fails for resp_id 2 and 5
-#' }
+#' try(expect_base(q1b, q1a %in% 1, data = my_survey)) # Fails for resp_id 2 and 5
 #'
 #' # Check that q2b has a value if and only if q2a < 50
 #' expect_base(q2b, q2a < 50, data = my_survey)
@@ -40,7 +38,8 @@ NULL
 #' @param base condition for missing check
 #' @param missing_valid allow missing values for records meeting the condition.
 #'   This allows 'one way' base checks. This is `FALSE` by default.
-expect_base <- function(var, base, miss = getOption("testdat.miss"), missing_valid = FALSE, data = get_testdata()) {
+expect_base <- function(var, base, miss = getOption("testdat.miss"),
+                        missing_valid = FALSE, data = get_testdata()) {
   act <- quasi_label(enquo(data))
 
   act$var_desc <- as_label(ensym(var))
@@ -99,4 +98,3 @@ expect_cond <- function(cond1, cond2, data = get_testdata()) {
 
   invisible(act$result)
 }
-