+++ R/utils.R
@@ -4,3 +4,8 @@ as_char_scipen <- function(x) {
   on.exit(options(old))
   as.character(x)
 }
+
+# TODO: remove when minimum R version >= 3.5
+if (getRversion() < 3.5) {
+  isFALSE <- function(x) is.logical(x) && length(x) == 1L && !is.na(x) && !x
+}