+++ scratch/test-folder.R
@@ -1,4 +1,0 @@
-library(testdat)
-
-x <- test_dir("example/auto/")
-output_results_excel(x, "example/Excel output test.xlsx")