+++ tests/testthat/test-expect_custom.R
@@ -1,5 +1,3 @@
-context("expect_custom")
-
 test_that("expect warns if no `failure_message`", {
   expect_warning(expect_custom(TRUE), "missing, with no default")
 })