+++ tests/testthat/test-expect_exclusive.R
@@ -1,5 +1,4 @@
 suppressWarnings(library(dplyr))
-context("expect_exclusive")
 
 dat <- tribble(
   ~a1, ~a2, ~a3, ~a4, ~a5, ~a6, ~a7, ~a8, ~a9, ~a10,