+++ NEWS.md
@@ -1,3 +1,13 @@
+# testdat (development version)
+
+* Soft deprecated `context_data()` (#43). `context_data()` is just a wrapper for `set_testdata()`, which has a much more intuitive name.
+
+* Soft deprecated `expect_similar()` (#18). It was a silly way of comparing data frames and we're better off making something new.
+
+* `chk_blank()` performs checks slightly differently and is much faster as a result (#46).
+
+* `chk_ascii()` was returning `FALSE` if it detected non-ASCII printable characters anywhere in the input vector. It now checks each element of the vector individually.
+
 # testdat 0.2.0
 
 In addition to minor updates and bug fixes, this release does three main things: